+++ samples/MicrosoftExchange/Exchange-Online-Persistent-Shell:DistLists-Management/PowershellScripts/EXOMailUser/createScript.ps1
@@ -1,0 +1,8 @@
+param(
+    [string]$name,
+    [string]$guid,
+    [string[]]$memberOf
+)
+
+
+