+++ samples/MicrosoftExchange/Exchange-Online-Persistent-Shell:DistLists-Management/PowershellScripts/EXOMailUser/deleteScript.ps1
@@ -1,0 +1,6 @@
+param(
+    [string]$guid
+)
+
+
+